Added Bones for Contains/Prefix
It seems as if the different searches are going to potentially be more
complicated than originally thought by our team. For anybody reading
this, have you had any luck with how to search based on specific
criteria? It appears the project may need to be rearranged a bit to
allow for more flexible searches.